feat(chat): add promote and demote participant as admin

+++ app/Http/Controllers/ChatController.php
@@ -9,32 +9,50 @@ use App\Http\Resources\UserResource;
 use App\Models\Chat;
 use App\Models\User;
 use App\Services\ChatService;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
 use Illuminate\Validation\ValidationException;
 
 class ChatController extends Controller
 {
-    use ValidatesRequests;
+    use AuthorizesRequests, ValidatesRequests;
 
     public function __construct(private readonly ChatService $chatService) {}
 
-    public function index(Request $request)
+    /**
+     * @throws AuthorizationException
+     */
+    public function index(Request $request): AnonymousResourceCollection
     {
-        return $this->chatService->getAllPaginated($request);
+        $this->authorize('view-any', Chat::class);
+
+        return ChatResource::collection($this->chatService->getAllPaginated($request));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function store(StoreChatRequest $request): ChatResource
     {
+        $this->authorize('create', Chat::class);
+
         $chat = $this->chatService->create($request);
 
         return ChatResource::make($chat);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(Chat $chat, UpdateChatRequest $request): JsonResponse
     {
+        $this->authorize('update', $chat);
+
         $chat = $this->chatService->update($chat, $request);
 
         return ChatResource::make($chat)->response()->setStatusCode(200);
@@ -42,9 +60,12 @@ class ChatController extends Controller
 
     /**
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function addParticipant(Chat $chat, User $user): UserResource
     {
+        $this->authorize('group-admin', $chat);
+
         $chat = $this->chatService->addParticipant($chat, $user);
 
         return UserResource::make($chat);
@@ -52,16 +73,50 @@ class ChatController extends Controller
 
     /**
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function removeParticipant(Chat $chat, User $user): UserResource
     {
+        $this->authorize('group-admin', $chat);
+
         $chat = $this->chatService->removeParticipant($chat, $user);
 
         return UserResource::make($chat);
     }
 
+    /**
+     * @throws ValidationException
+     * @throws AuthorizationException
+     */
+    public function makeAsAdmin(Chat $chat, User $user): UserResource
+    {
+        $this->authorize('group-admin', $chat);
+
+        $chat = $this->chatService->makeAsAdmin($chat, $user);
+
+        return UserResource::make($chat);
+    }
+
+    /**
+     * @throws ValidationException
+     * @throws AuthorizationException
+     */
+    public function dismissAsAdmin(Chat $chat, User $user): UserResource
+    {
+        $this->authorize('group-admin', $chat);
+
+        $chat = $this->chatService->dismissAsAdmin($chat, $user);
+
+        return UserResource::make($chat);
+    }
+
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Chat $chat): Response
     {
+        $this->authorize('delete', $chat);
+
         $this->chatService->delete($chat);
 
         return response()->noContent();
